+++ workstation/update.sh
@@ -10,18 +10,13 @@ UPDATE_DUMP="$KIRA_DUMP/kiraup"
 MAX_FAILS=3
 
 UPDATE_FAILS=$(globGet UPDATE_FAIL_COUNTER)
+SETUP_START_DT=$(globGet SETUP_START_DT)
+IS_WSL=$(isSubStr "$(uname -a)" "microsoft-standard-WSL")
 
 if (! $(isNaturalNumber $UPDATE_FAILS)) ; then
     UPDATE_FAILS=0
     globSet UPDATE_FAIL_COUNTER $UPDATE_FAILS
 fi 
-
-SETUP_START_DT=$(globGet SETUP_START_DT)
-# marks if system was rebooted before tools setup started (this is required in case of docker deamon malfunction)
-SETUP_REBOOT=$(globGet SETUP_REBOOT)
-# do not reboot if KM is run in WSL mode
-IS_WSL=$(isSubStr "$(uname -a)" "microsoft-standard-WSL")
-[ "${IS_WSL,,}" == "true" ] && SETUP_REBOOT="true"
 
 if [[ $UPDATE_FAILS -ge $MAX_FAILS ]] ; then
     echoErr "ERROR: Stopping update service for error..."
@@ -33,8 +28,7 @@ if [[ $UPDATE_FAILS -ge $MAX_FAILS ]] ; then
         cat $KIRA_LOGS/kirascan.log > "$KIRA_DUMP/kirascan-done.log.txt" || echoErr "ERROR: Failed to dump kira scan service log"
     fi
     echoErr "Press 'Ctrl+c' to exit then type 'kira' to enter infra manager"
-    sleep 5
-    systemctl stop kiraup
+    sleep 5 && systemctl stop kiraup
     exit 1
 fi
 
@@ -47,7 +41,7 @@ echoWarn "|     FAILS COUNTER: $UPDATE_FAILS"
 echoWarn "|         MAX FAILS: $MAX_FAILS"
 echoWarn "| SETUP START DTATE: $SETUP_START_DT"
 echoWarn "|   SETUP END DTATE: $SETUP_END_DT"
-echoWarn "|      SETUP REBOOT: $SETUP_REBOOT"
+echoWarn "|      SETUP REBOOT: $(globGet SETUP_REBOOT)"
 echoWarn "------------------------------------------------"
 
 mkdir -p $UPDATE_DUMP
@@ -56,12 +50,12 @@ mkdir -p $UPDATE_DUMP
 
 if [ "$(globGet "ESSENAILS_UPDATED_$KIRA_SETUP_VER")" != "true" ]; then
     echoInfo "INFO: Installing essential tools and dependecies"
-
-    if [ -z "$SETUP_REBOOT" ] ; then
+ 
+    if [ "$(globGet SETUP_REBOOT)" != "done" ] ; then
         echoInfo "INFO: Reboot is required before tools setup can continue..." && sleep 3
         echoErr "Reconnect to your machine after restart and type 'kira' in the console to continue"
         globSet SETUP_REBOOT "done"
-        reboot
+        [ "${IS_WSL,,}" != "true" ] && reboot
         exit 0
     else
         echoInfo "INFO: Tools setup reboot was already performed, setup will continue..."
@@ -118,11 +112,11 @@ else
     echoInfo "INFO: Environment cleanup was already executed"
 fi
 
-if [ -z "$SETUP_REBOOT" ] && [ "${IS_WSL,,}" != "true" ] ; then
+if [ "$(globGet SETUP_REBOOT)" != "done" ] ; then
     echoInfo "INFO: Reboot is required before setup can continue..." && sleep 3
     echoErr "Reconnect to your machine after restart and type 'kira' in the console to continue"
     globSet SETUP_REBOOT "done"
-    reboot
+    [ "${IS_WSL,,}" != "true" ] && reboot
     exit 0
 else
     echoInfo "INFO: Reboot was already performed, setup will continue..."
@@ -148,8 +142,7 @@ if [ "$(globGet "CONTAINERS_UPDATED_$KIRA_SETUP_VER")" != "true" ]; then
         globSet "CLEANUPS_UPDATED_$KIRA_SETUP_VER" "false"
         globSet "ESSENAILS_UPDATED_$KIRA_SETUP_VER" "false"
         globSet UPDATE_FAIL_COUNTER $(($UPDATE_FAILS + 1))
-        sleep 5
-        reboot
+        sleep 5 && [ "${IS_WSL,,}" != "true" ] && reboot
     fi
 else
     echoInfo "INFO: Docker containers were already updated"
@@ -165,9 +158,10 @@ if [ "${UPDATE_DONE,,}" == "true" ] ; then
         cat $KIRA_LOGS/kiraup.log > "$KIRA_DUMP/kiraup-done.log.txt" || echoErr "ERROR: Failed to dump kira update service log"
         cat $KIRA_LOGS/kirascan.log > "$KIRA_DUMP/kirascan-done.log.txt" || echoErr "ERROR: Failed to dump kira scan service log"
     fi
-    set +x
+    
     globSet SETUP_END_DT "$(date +'%Y-%m-%d %H:%M:%S')"
     globSet UPDATE_DONE "true"
+    set +x
     echoInfo "Press 'Ctrl+c' to exit then type 'kira' to enter infra manager"
     sleep 5
     systemctl stop kiraup 
@@ -181,5 +175,5 @@ echoWarn "------------------------------------------------"
 echoWarn "| FINISHED: UPDATE SCRIPT $KIRA_SETUP_VER"
 echoWarn "|  ELAPSED: $(($(date -u +%s) - $SCRIPT_START_TIME)) seconds"
 echoWarn "------------------------------------------------"
-
 sleep 10
+set -x
